requirements.txt
aiohttp==3.8.1
Django==3.2.13
django_jsonfield==1.4.1
djangorestframework==3.13.1
requests==2.27.1
